rename AsciiSerializer to StringSerializer
The current AsciiSerializer is actually the default string serializer
and should also be named as such to avoid confusion. Speaking of ASCII
serialization is also strictly speaking no longer correct, since
strings can contain UTF-8 characters.
