+++ tests/CMakeLists.txt
@@ -15,10 +15,10 @@ target_dots_model(${TARGET_NAME}
 )
 target_sources(${TARGET_NAME}
     PRIVATE
-        src/io/serialization/TestAsciiSerializer.cpp
         src/io/serialization/TestCborSerializer.cpp
         src/io/serialization/TestExperimentalCborSerializer.cpp
         src/io/serialization/TestJsonSerializer.cpp
+        src/io/serialization/TestStringSerializer.cpp
 
         src/TestDigest.cpp
         src/TestAsciiSerialization.cpp
