+++ tests/src/io/serialization/data/string_serialization_data.h
